+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changelog
 Next Version
 ------------
 
+* Support ``fromYAML`` classmethod for all Munch subclasses (PR [#52](https://github.com/Infinidat/munch/pull/52) fixes [#34](https://github.com/Infinidat/munch/issues/34)
+
 2.4.0 (2019-10-29)
 ------------------